editoast: add ToToken ModelV2 definition ExistsImpl
